--- src/main/java/com/ecovision/ecovision/controller/AdminController.java
@@ -1,18 +1,0 @@
-package com.ecovision.ecovision.controller;
-
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.ResponseBody;
-
-@Controller
-@ResponseBody
-public class AdminController {
-
-    @GetMapping("/admin")
-    public String adminP() {
-
-        return "Admin Controller";
-    }
-
-}
