+++ src/main/java/com/ecovision/ecovision/dto/UserRequestDto.java
@@ -1,0 +1,2 @@
+package com.ecovision.ecovision.dto;public class UserRequestDto {
+}
