+++ src/main/java/com/ecovision/ecovision/dto/UserResponseDto.java
@@ -1,0 +1,2 @@
+package com.ecovision.ecovision.dto;public class UserResponseDto {
+}
